Set the min Windows target to 8.1 and fix various MSVC compiler warnings. Win 8.1 is required for some ASIO networking functionality, DPI support in SDL and other stuff. Also make the Windows target consistent across projects instead of defaulting to whatever each individual library decides. Remove also the Windows XP info doc since this is no longer a target.

+++ vulkan_gl/RawBuffer.cpp
@@ -109,7 +109,7 @@ bool RawBuffer::init(
     vkFuncs.vkGetBufferMemoryRequirements(vkDevice, mVkBuffer, &memReqs);
 
     // Try to do a device memory alloc
-    deviceMemMgr.alloc(memReqs.size, memReqs.memoryTypeBits, memReqs.alignment, allocMode, mDeviceMemAlloc);
+    deviceMemMgr.alloc(memReqs.size, memReqs.memoryTypeBits, (uint32_t) memReqs.alignment, allocMode, mDeviceMemAlloc);
 
     // If that fails then bail out and return false for failure
     if (mDeviceMemAlloc.size <= 0)
